Add move-construction support.
A new concept "move_constructible" was added similar to
"copy_constructible" concept. It is a composition of "destructible" and
"constructible(_self(self&&)".

However, to prevent ambiguities, in case "copy_constructible" and
"move_constructible" are both in concept, the latter is implemented
with some other internals which are used when deducing the correct
constructor.